* Add hour when "hour" resolution is used * Change date format to DAY/MONTH/YEAR * Start Year dropdown from the year of the first timestamp
