+++ admin-front/src/components/TimeSeries/functions/index.ts
@@ -1,0 +1,2 @@
+export * from './getDateFormat';
+export * from './getTimeSeriesFilteredByOccurrence';
