+++ admin-front/src/components/TimeSeries/hooks/useGetTimeSeriesData.tsx
@@ -1,25 +1,24 @@
 import { useEffect, useState } from 'react';
 import moment from 'moment';
 import useGetDataFromAPI from 'shared/hooks/useGetDataFromAPI';
-import { getTimeSeriesFilteredByOccurrence } from '../functions/getTimeSeriesFilteredByOccurrence';
-import { Month, Resolution } from '../types';
-
-export interface TimeSeriesRecord {
-  timestamp: number;
-  value: number;
-}
+import { getTimeSeriesFilteredByOccurrence, getDateFormat } from '../functions';
+import {
+  Month,
+  Resolution,
+  TimeSeriesRecord,
+  TimeSeriesType,
+} from '../utility';
 
 interface TimeSeriesData {
   timeStamps: string[];
   timeStampValues: number[];
+  firstYearTimeSeriesRecord?: TimeSeriesRecord;
 }
 
 const TimeSeriesPaths = {
   nftCount: '/analytics/sales/nftCount/timeseries',
   priceVolume: '/analytics/sales/priceVolume/timeseries',
 };
-
-export type TimeSeriesType = 'nftCount' | 'priceVolume';
 
 interface UseGetTimeSeriesDataProps {
   timeSeriesType: TimeSeriesType;
@@ -52,7 +51,8 @@ const UseGetTimeSeriesData = ({
 
       setTimeStamps(
         timeSeriesFilteredByOccurrence.map((record: TimeSeriesRecord) => {
-          return moment.unix(record.timestamp).format('MM/DD/YYYY');
+          const dateFormat = getDateFormat({ resolution, month });
+          return moment.unix(record.timestamp).format(dateFormat);
         }),
       );
 
@@ -67,6 +67,7 @@ const UseGetTimeSeriesData = ({
   return {
     timeStamps,
     timeStampValues,
+    firstYearTimeSeriesRecord: fetchedTimeSeries?.[0],
   };
 };
