--- admin-front/src/components/TimeSeries/types.ts
@@ -1,15 +1,0 @@
-export type Resolution = 'hour' | 'day' | 'week' | 'month';
-export type Month =
-  | 'All'
-  | 'January'
-  | 'February'
-  | 'March'
-  | 'April'
-  | 'May'
-  | 'June'
-  | 'July'
-  | 'August'
-  | 'September'
-  | 'October'
-  | 'November'
-  | 'December';
