+++ admin-front/src/components/TimeSeries/utility/constants.ts
@@ -1,0 +1,66 @@
+import moment from 'moment/moment';
+
+export const getYears = (startYear: number) => {
+  const years = [];
+
+  for (let year = startYear; year <= moment().year(); year++) {
+    years.push({ value: year, label: year });
+  }
+
+  return years;
+};
+
+export const months = [
+  {
+    value: 'All',
+    label: 'All months',
+  },
+  {
+    value: 'January',
+    label: 'January',
+  },
+  {
+    value: 'February',
+    label: 'February',
+  },
+  {
+    value: 'March',
+    label: 'March',
+  },
+  {
+    value: 'April',
+    label: 'April',
+  },
+  {
+    value: 'May',
+    label: 'May',
+  },
+  {
+    value: 'June',
+    label: 'June',
+  },
+  {
+    value: 'July',
+    label: 'July',
+  },
+  {
+    value: 'August',
+    label: 'August',
+  },
+  {
+    value: 'September',
+    label: 'September',
+  },
+  {
+    value: 'October',
+    label: 'October',
+  },
+  {
+    value: 'November',
+    label: 'November',
+  },
+  {
+    value: 'December',
+    label: 'December',
+  },
+];
